+++ src/main/java/implementations/Report1RowMapper.java
@@ -3,7 +3,6 @@ package implementations;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 
-import java.math.BigInteger;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
@@ -15,7 +14,7 @@ public class Report1RowMapper implements RowMapper<Report1> {
     @Override
     public Report1 mapRow(ResultSet row, int rowNum) throws SQLException {
         Report1 rep1 = new Report1(jdbcTemplate);
-        rep1.setId(BigInteger.valueOf(row.getInt("ID")));
+        rep1.setId(row.getInt("ID"));
         rep1.setName(row.getString("NAME"));
         rep1.setSum(row.getDouble("SUMMA"));
         return rep1;
