+++ src/main/java/interfaces/BudgetType.java
@@ -1,10 +1,9 @@
 package interfaces;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 
 public interface BudgetType extends Entity {
-    BigInteger getBudgetTypeId();
+    Integer getBudgetTypeId();
     Integer getGroupId();
     String getName();
     Boolean getRequired();
