+++ src/main/java/interfaces/Person.java
@@ -1,10 +1,9 @@
 package interfaces;
 
-import java.math.BigInteger;
 import java.util.Date;
 
 public interface Person extends Entity {
-    BigInteger getPersonId();
+    Integer getPersonId();
     String getLogin();
     String getPass();
     String getEmail();
